requirements.txt
django
djangorestframework
tzdata